backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
aiofiles
openai
python-dotenv
xhtml2pdf
